Encapsulate GraphQL
